backend/requirements.txt
fastapi
uvicorn[standard]
websockets
python-multipart
pydantic
pydantic-settings
httpx
faster-whisper
numpy
soundfile
piper-tts